bitstring_trees

Module iter

Iterators over bit string prefixes

Structs

IterBetween

Generate smallest set of prefixes covering values between other prefixes

IterInclusive

Generate the smallest (ordered) list of prefixes covering first..=last

Functions

iter_between

Generate smallest set of prefixes covering values between start and end

iter_inclusive

Generate the smallest (ordered) list of prefixes covering first..=last