bitstring_trees::iter

Function iter_between

pub fn iter_between<K>(after: Option<K>, before: Option<K>) -> IterBetween<K> 
where
    K: BitString + Clone,

Generate smallest set of prefixes covering values between start and end

Pass None to cover all values before or after a prefix, or simply all values.