bitstring_trees
Module tree

generic map of bit strings prefixes to values

This is a very generic abstraction and therefore not easy to use.

Look for other containers in this crate that offer specific use cases.

Structs

DefaultCompare

Use Eq for LeafValueComparer

IterInOrder

Iterate over node of tree depth-first in-order

IterLeaf

Iterate over nodes and leaf values of tree in-order

IterLeafFull

Iterate over keys and mutable leaf values and uncovered keys of tree in-order

IterMutBorrowedInOrder

Iterate over keys and mutable values of tree depth-first in-order

IterMutBorrowedLeaf

Iterate over keys and mutable leaf values of tree in-order

IterMutBorrowedLeafFull

Iterate over keys and mutable leaf values and uncovered keys of tree in-order

IterMutBorrowedPostOrder

Iterate over keys and mutable values of tree depth-first post-order

IterMutBorrowedPreOrder

Iterate over keys and mutable values of tree depth-first pre-order

IterMutOwnedInOrder

Iterate over keys and mutable values of tree depth-first in-order

IterMutOwnedLeaf

Iterate over keys and mutable leaf values of tree in-order

IterMutOwnedLeafFull

Iterate over keys and mutable leaf values and uncovered keys of tree in-order

IterMutOwnedPostOrder

Iterate over keys and mutable values of tree depth-first post-order

IterMutOwnedPreOrder

Iterate over keys and mutable values of tree depth-first pre-order

IterMutPath

Iterate over keys and mutable values of tree that are a prefix of target key

IterPath

Iterate over all nodes that are a prefix of target key

IterPostOrder

Iterate over node of tree depth-first post-order

IterPreOrder

Iterate over node of tree depth-first pre-order

IterWalkMutBorrowedPath

Iterate over all nodes that are a prefix of target key in a WalkMutBorrowed stack

IterWalkMutOwnedPath

Iterate over all nodes that are a prefix of target key in a WalkMutOwned stack

MutPath

Iterate over all nodes that are a prefix of target key

NoEqual

Define no leaf values to be equal for LeafValueComparer

Node

Node in tree

Tree

Tree is a binary tree with path-shortening.

Walk

Walk tree

WalkMutBorrowed

Walk borrowed mutable tree up and down

WalkMutBorrowedPath

Iterate over all nodes that are a prefix of target key in a WalkMutBorrowed stack

WalkMutOwned

Walk owned mutable tree up and down

WalkMutOwnedPath

Iterate over all nodes that are a prefix of target key in a WalkMutOwned stack

Enums

InsertPosition

Result of key lookup in tree

InsertPositionWith

Result of finding position to insert a target key

WalkedDirection

Remember which path was taken to reach this node from the parent in WalkMutOwned and WalkMutBorrowed.

Traits

LeafValueComparer

Define how to compare leaf values in tree

TreeProperties

Define Tree behavior